+++ secureByDesign/netflix2/csharp/Netflix2.tests/DvdDaoTests.cs
@@ -1,0 +1,20 @@
+﻿using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using Netflix2.repository;
+using NFluent;
+using NUnit.Framework;
+
+namespace Netflix2.tests
+{
+    [TestFixture]
+    public class DvdDaoTests
+    {
+        [Test]
+        public void GetIdentifiers_Ok()
+        {
+            var actual = DvdDao.GetIdentifiers();
+            Check.That(actual.Count()).IsEqualTo(8);
+        }
+    }
+}
