+++ secureByDesign/netflix2/csharp/Netflix2/Netflix2.csproj
@@ -65,6 +65,8 @@
   <ItemGroup>
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="repository\DB.cs" />
+    <Compile Include="repository\DvdDao.cs" />
     <Compile Include="repository\UserDao.cs" />
   </ItemGroup>
   <ItemGroup>
